+++ Solana/crypto-ticket/programs/crypto-ticket/src/instructions/cleanup.rs
@@ -22,7 +22,7 @@ pub fn cleanup_participants_chunk(ctx: Context<CleanupParticipantsChunk>) -> Res
     emit!(ChunkCleanedEvent {
         ticket_id: chunk.ticket_id,
         chunk_index: chunk.chunk_index,
-        timestamp: Clock::get()?.unix_timestamp,
+        timestamp: Clock::get()?.unix_timestamp as u64,
     });
 
     Ok(())
@@ -53,8 +53,8 @@ pub fn batch_cleanup_chunks(ctx: Context<BatchCleanupChunks>, chunk_indexes: Vec
 
     emit!(BatchCleanupEvent {
         ticket_id: ticket_account.ticket_id,
-        chunks_cleaned: chunk_indexes.len() as i64,
-        timestamp: Clock::get()?.unix_timestamp,
+        chunks_cleaned: chunk_indexes.len() as u64,
+        timestamp: Clock::get()?.unix_timestamp as u64,
     });
 
     Ok(())
